refactor : state, effect 훅으로 분리
