+++ admin/src/components/EventManagement/FCFSModal.tsx
@@ -1,76 +1,33 @@
-import { useEffect, useState } from 'react';
 import { Calendar } from '../ui/calendar';
 import { TimePicker } from '@/components/ui/datetime-picker';
 import Modal from '@/components/common/Modal';
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
 import { ko } from 'date-fns/locale';
-
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover';
 import { CalendarIcon } from 'lucide-react';
-import { format, setHours, setMinutes } from 'date-fns';
+import useFcFsRequest from '@/hooks/useFcFsRequest';
 import { FcFsRequest } from '@/types/eventType';
+import { format } from 'date-fns';
+
 interface FCFSModalProps {
   isOpen: boolean;
   handleClose: () => void;
   handleSave: (body: FcFsRequest) => void;
 }
 
-const initRequest: FcFsRequest = {
-  startDate: format(new Date(), 'yyyy-MM-dd'),
-  endDate: format(new Date(), 'yyyy-MM-dd'),
-  startTime: format(new Date(), 'HH:mm:ss'),
-};
-
 const FCFSModal = ({ isOpen, handleClose, handleSave }: FCFSModalProps) => {
-  const [fcfsRequest, setFcfsRequest] = useState<FcFsRequest>(initRequest);
-  const [endFCFSTime, setEndFCFSTime] = useState<Date | undefined>(undefined);
-
-  useEffect(() => {
-    if (fcfsRequest.startTime) {
-      const startFCFSTime = new Date(
-        `${fcfsRequest.startDate}T${fcfsRequest.startTime}`
-      );
-      const twoHoursLater = new Date(startFCFSTime);
-      twoHoursLater.setHours(startFCFSTime.getHours() + 2);
-
-      const endOfDay = setMinutes(setHours(startFCFSTime, 23), 59);
-      const finalEndTime = twoHoursLater > endOfDay ? endOfDay : twoHoursLater;
-
-      setEndFCFSTime(finalEndTime);
-    }
-  }, [fcfsRequest.startTime, fcfsRequest.startDate]);
-
-  const handleStartDateChange = (date: Date | undefined) => {
-    if (date) {
-      setFcfsRequest((prevRequest) => ({
-        ...prevRequest,
-        startDate: format(date, 'yyyy-MM-dd'),
-      }));
-    }
-  };
-
-  const handleEndDateChange = (date: Date | undefined) => {
-    if (date) {
-      setFcfsRequest((prevRequest) => ({
-        ...prevRequest,
-        endDate: format(date, 'yyyy-MM-dd'),
-      }));
-    }
-  };
-
-  const handleStartTimeChange = (date: Date | undefined) => {
-    if (date) {
-      setFcfsRequest((prevRequest) => ({
-        ...prevRequest,
-        startTime: format(date, 'HH:mm:ss'),
-      }));
-    }
-  };
+  const {
+    fcfsRequest,
+    endFCFSTime,
+    handleStartDateChange,
+    handleEndDateChange,
+    handleStartTimeChange,
+  } = useFcFsRequest();
 
   if (!isOpen) return null;
 
@@ -78,7 +35,7 @@ const FCFSModal = ({ isOpen, handleClose, handleSave }: FCFSModalProps) => {
     <div className='top-[270px] left-[780px] flex fixed inset-0 w-fit h-fit flex-row'>
       <Modal
         handleCloseClick={handleClose}
-        handleButtonClick={() => handleSave(initRequest)}
+        handleButtonClick={() => handleSave(fcfsRequest)}
       >
         <div className='flex flex-row mx-3'>
           <div className='space-y-3 space-x-2 flex flex-col h-[180px] mt-1'>
